+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - Support for OpenSeadragon 5
+- `OMEZarrTileSource.DUMMY_XHR`; tile jobs now finish and fail with a `null` request
 
 ## [0.2.0] - 2026-05-25